Fix stripping prefix when nesting at `/`
Fixes https://github.com/tokio-rs/axum/issues/88
