+++ src/routing.rs
@@ -958,15 +958,17 @@ where
 
 fn strip_prefix(uri: &Uri, prefix: &str) -> Uri {
     let path_and_query = if let Some(path_and_query) = uri.path_and_query() {
-        let mut new_path = if let Some(path) = path_and_query.path().strip_prefix(prefix) {
+        let new_path = if let Some(path) = path_and_query.path().strip_prefix(prefix) {
             path
         } else {
             path_and_query.path()
         };
 
-        if new_path.is_empty() {
-            new_path = "/";
-        }
+        let new_path = if new_path.starts_with('/') {
+            Cow::Borrowed(new_path)
+        } else {
+            Cow::Owned(format!("/{}", new_path))
+        };
 
         if let Some(query) = path_and_query.query() {
             Some(
