+++ src/tests.rs
@@ -1,10 +1,10 @@
 use crate::{
-    extract::RequestParts, handler::on, prelude::*, response::IntoResponse, routing::MethodFilter,
-    service,
+    extract::RequestParts, handler::on, prelude::*, response::IntoResponse, routing::nest,
+    routing::MethodFilter, service,
 };
 use bytes::Bytes;
 use futures_util::future::Ready;
-use http::{header::AUTHORIZATION, Request, Response, StatusCode};
+use http::{header::AUTHORIZATION, Request, Response, StatusCode, Uri};
 use hyper::{Body, Server};
 use serde::Deserialize;
 use serde_json::json;
@@ -16,6 +16,8 @@ use std::{
 };
 use tower::{make::Shared, service_fn, BoxError, Service, ServiceBuilder};
 use tower_http::{compression::CompressionLayer, trace::TraceLayer};
+
+mod nest;
 
 #[tokio::test]
 async fn hello_world() {
@@ -522,72 +524,6 @@ async fn layer_on_whole_router() {
 }
 
 #[tokio::test]
-async fn disjunction() {
-    let api_routes = route(
-        "/users",
-        get(|| async { "users#index" }).post(|| async { "users#create" }),
-    )
-    .route(
-        "/users/:id",
-        get(|params: extract::UrlParamsMap| async move {
-            format!(
-                "{}: users#show ({})",
-                params.get("version").unwrap(),
-                params.get("id").unwrap()
-            )
-        }),
-    )
-    .route(
-        "/games/:id",
-        get(|params: extract::UrlParamsMap| async move {
-            format!(
-                "{}: games#show ({})",
-                params.get("version").unwrap(),
-                params.get("id").unwrap()
-            )
-        }),
-    );
-
-    let app = route("/", get(|| async { "hi" })).nest("/:version/api", api_routes);
-
-    let addr = run_in_background(app).await;
-
-    let client = reqwest::Client::new();
-
-    let res = client
-        .get(format!("http://{}/", addr))
-        .send()
-        .await
-        .unwrap();
-    assert_eq!(res.status(), StatusCode::OK);
-    assert_eq!(res.text().await.unwrap(), "hi");
-
-    let res = client
-        .get(format!("http://{}/v0/api/users", addr))
-        .send()
-        .await
-        .unwrap();
-    assert_eq!(res.status(), StatusCode::OK);
-    assert_eq!(res.text().await.unwrap(), "users#index");
-
-    let res = client
-        .get(format!("http://{}/v0/api/users/123", addr))
-        .send()
-        .await
-        .unwrap();
-    assert_eq!(res.status(), StatusCode::OK);
-    assert_eq!(res.text().await.unwrap(), "v0: users#show (123)");
-
-    let res = client
-        .get(format!("http://{}/v0/api/games/123", addr))
-        .send()
-        .await
-        .unwrap();
-    assert_eq!(res.status(), StatusCode::OK);
-    assert_eq!(res.text().await.unwrap(), "v0: games#show (123)");
-}
-
-#[tokio::test]
 async fn typed_header() {
     use extract::TypedHeader;
     async fn handle(TypedHeader(user_agent): TypedHeader<headers::UserAgent>) -> impl IntoResponse {
@@ -710,33 +646,6 @@ async fn wrong_method_handler() {
 
     let res = client
         .get(format!("http://{}/bar", addr))
-        .send()
-        .await
-        .unwrap();
-    assert_eq!(res.status(), StatusCode::NOT_FOUND);
-}
-
-#[tokio::test]
-async fn wrong_method_nest() {
-    let nested_app = route("/", get(|| async {}));
-    let app = crate::routing::nest("/", nested_app);
-
-    let addr = run_in_background(app).await;
-
-    let client = reqwest::Client::new();
-
-    let res = client.get(format!("http://{}", addr)).send().await.unwrap();
-    assert_eq!(res.status(), StatusCode::OK);
-
-    let res = client
-        .post(format!("http://{}", addr))
-        .send()
-        .await
-        .unwrap();
-    assert_eq!(res.status(), StatusCode::METHOD_NOT_ALLOWED);
-
-    let res = client
-        .patch(format!("http://{}/foo", addr))
         .send()
         .await
         .unwrap();
